refactor: harmonize filenames and move borrow utils up to llamelend level
