+++ apps/main/src/llamalend/features/borrow/components/BorrowActionInfoAccordion.tsx
@@ -1,5 +1,5 @@
+import { formatPercent } from '@/llamalend/format.utils'
 import type { NetworkDict } from '@/llamalend/llamalend.types'
-import { formatPercent } from '@/llamalend/utils'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import Box from '@mui/material/Box'
 import Stack from '@mui/material/Stack'
@@ -12,13 +12,13 @@ import { GearIcon } from '@ui-kit/shared/icons/GearIcon'
 import { Accordion } from '@ui-kit/shared/ui/Accordion'
 import ActionInfo from '@ui-kit/shared/ui/ActionInfo'
 import { getHealthValueColor } from '../../market-position-details/utils'
-import { type BorrowForm, type BorrowFormQueryParams, type Token } from '../borrow.types'
 import { useLoanToValue } from '../hooks/useLoanToValue'
 import { useMarketRates } from '../queries/borrow-apy.query'
 import { useBorrowBands } from '../queries/borrow-bands.query'
 import { useBorrowEstimateGas } from '../queries/borrow-gas-estimate.query'
 import { useBorrowHealth } from '../queries/borrow-health.query'
 import { useBorrowPrices } from '../queries/borrow-prices.query'
+import { type BorrowForm, type BorrowFormQueryParams, type Token } from '../types'
 import { BorrowLeverageActionInfos } from './BorrowLeverageActionInfos'
 
 /**
