+++ apps/main/src/llamalend/features/borrow/components/BorrowTabContents.tsx
@@ -1,6 +1,7 @@
 import { useCallback, useEffect } from 'react'
 import type { ReactNode } from 'react'
 import { FormProvider } from 'react-hook-form'
+import { hasLeverage } from '@/llamalend/llama.utils'
 import type { LlamaMarketTemplate, NetworkDict } from '@/llamalend/llamalend.types'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import Box from '@mui/material/Box'
@@ -10,9 +11,8 @@ import Stack from '@mui/material/Stack'
 import { useBorrowPreset } from '@ui-kit/hooks/useLocalStorage'
 import { t } from '@ui-kit/lib/i18n'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { BorrowPreset, type OnBorrowFormUpdate } from '../borrow.types'
-import { hasLeverage } from '../llama.util'
-import { setValueOptions } from '../llama.util'
+import { setValueOptions } from '../react-form.utils'
+import { BorrowPreset, type OnBorrowFormUpdate } from '../types'
 import { useBorrowForm } from '../useBorrowForm'
 import { AdvancedBorrowOptions } from './AdvancedBorrowOptions'
 import { BorrowActionInfoAccordion } from './BorrowActionInfoAccordion'
