+++ apps/main/src/llamalend/features/borrow/components/LeverageInput.tsx
@@ -9,8 +9,8 @@ import { t } from '@ui-kit/lib/i18n'
 import ActionInfo from '@ui-kit/shared/ui/ActionInfo'
 import { WithSkeleton } from '@ui-kit/shared/ui/WithSkeleton'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import type { BorrowForm, BorrowFormQueryParams } from '../borrow.types'
 import { useBorrowExpectedCollateral } from '../queries/borrow-expected-collateral.query'
+import type { BorrowForm, BorrowFormQueryParams } from '../types'
 
 const { Spacing } = SizesAndSpaces
