+++ apps/main/src/llamalend/features/borrow/components/LiquidationRangeSlider.tsx
@@ -7,7 +7,7 @@ import { formatNumber } from '@ui/utils'
 import { t } from '@ui-kit/lib/i18n'
 import { NumericTextField } from '@ui-kit/shared/ui/NumericTextField'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { BORROW_PRESET_RANGES } from '../llama.util'
+import { BORROW_PRESET_RANGES } from '../constants'
 
 const { Spacing } = SizesAndSpaces
 const format = (value: number) => formatNumber(value, { style: 'currency', currency: 'USD' })
