+++ apps/main/src/llamalend/features/borrow/components/LoanPresetSelector.tsx
@@ -6,8 +6,8 @@ import ToggleButton from '@mui/material/ToggleButton'
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup'
 import { t } from '@ui-kit/lib/i18n'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { BorrowPreset } from '../borrow.types'
-import { BORROW_PRESET_RANGES } from '../llama.util'
+import { BORROW_PRESET_RANGES } from '../constants'
+import { BorrowPreset } from '../types'
 
 const PRESETS = {
   [BorrowPreset.Safe]: {
