+++ apps/main/src/llamalend/features/borrow/constants.ts
@@ -1,0 +1,7 @@
+import { BorrowPreset } from './types'
+
+export const BORROW_PRESET_RANGES = {
+  [BorrowPreset.Safe]: 50,
+  [BorrowPreset.MaxLtv]: 4,
+  [BorrowPreset.Custom]: 10,
+}
