+++ apps/main/src/llamalend/features/borrow/hooks/useMaxTokenValues.tsx
@@ -2,10 +2,10 @@ import { useEffect } from 'react'
 import type { UseFormReturn } from 'react-hook-form'
 import { type Address } from 'viem'
 import { useTokenBalance } from '@ui-kit/hooks/useTokenBalance'
-import type { BorrowForm, BorrowFormQueryParams } from '../borrow.types'
-import { setValueOptions } from '../llama.util'
 import { useMaxLeverage } from '../queries/borrow-max-leverage.query'
 import { useMaxBorrowReceive } from '../queries/borrow-max-receive.query'
+import { setValueOptions } from '../react-form.utils'
+import type { BorrowForm, BorrowFormQueryParams } from '../types'
 
 /**
  * Hook to fetch and set the maximum token values for collateral and debt in a borrow form.
