+++ apps/main/src/llamalend/features/borrow/queries/borrow-apy.query.ts
@@ -1,10 +1,10 @@
+import { getLlamaMarket } from '@/llamalend/llama.utils'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { LendMarketTemplate } from '@curvefi/llamalend-api/lib/lendMarkets'
 import { type FieldsOf } from '@ui-kit/lib'
 import type { PoolQuery } from '@ui-kit/lib/model'
 import { queryFactory, rootKeys } from '@ui-kit/lib/model'
 import { llamaApiValidationSuite } from '@ui-kit/lib/model/query/curve-api-validation'
-import { getLlamaMarket } from '../llama.util'
 import { maxBorrowReceiveKey } from './borrow-max-receive.query'
 
 type BorrowApyQuery = PoolQuery<IChainId>
