+++ apps/main/src/llamalend/features/borrow/queries/borrow-gas-estimate.query.ts
@@ -1,5 +1,6 @@
 import { useMemo } from 'react'
 import { ethAddress } from 'viem'
+import { getLlamaMarket } from '@/llamalend/llama.utils'
 import { type NetworkDict } from '@/llamalend/llamalend.types'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { LendMarketTemplate } from '@curvefi/llamalend-api/lib/lendMarkets'
@@ -7,8 +8,7 @@ import { type FieldsOf } from '@ui-kit/lib'
 import { queryFactory, rootKeys } from '@ui-kit/lib/model'
 import { calculateGas, useGasInfoAndUpdateLib } from '@ui-kit/lib/model/entities/gas-info'
 import { useTokenUsdRate } from '@ui-kit/lib/model/entities/token-usd-rate'
-import type { BorrowFormQuery } from '../borrow.types'
-import { getLlamaMarket } from '../llama.util'
+import type { BorrowFormQuery } from '../types'
 import { maxBorrowReceiveKey } from './borrow-max-receive.query'
 import { borrowQueryValidationSuite } from './borrow.validation'
