+++ apps/main/src/llamalend/features/borrow/queries/borrow-max-leverage.query.ts
@@ -1,11 +1,11 @@
 import { group } from 'vest'
+import { getLlamaMarket, hasLeverage } from '@/llamalend/llama.utils'
 import type { IChainId } from '@curvefi/api/lib/interfaces'
 import { MintMarketTemplate } from '@curvefi/llamalend-api/lib/mintMarkets'
 import { createValidationSuite, type FieldsOf } from '@ui-kit/lib'
 import { type PoolQuery, queryFactory, rootKeys } from '@ui-kit/lib/model'
 import { chainValidationGroup } from '@ui-kit/lib/model/query/chain-validation'
 import { llamaApiValidationGroup } from '@ui-kit/lib/model/query/curve-api-validation'
-import { hasLeverage, getLlamaMarket } from '../llama.util'
 import { validateRange } from './borrow.validation'
 
 type BorrowMaxLeverageQuery = PoolQuery<IChainId> & { range: number }
