+++ apps/main/src/llamalend/features/borrow/queries/borrow-oracle-price-band.query.ts
@@ -1,7 +1,7 @@
+import { getLlamaMarket } from '@/llamalend/llama.utils'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { type PoolParams, type PoolQuery, queryFactory, rootKeys } from '@ui-kit/lib/model'
 import { llamaApiValidationSuite } from '@ui-kit/lib/model/query/curve-api-validation'
-import { getLlamaMarket } from '../llama.util'
 
 export const { useQuery: useBorrowOraclePriceBand } = queryFactory({
   queryKey: ({ chainId, poolId }: PoolParams<IChainId>) =>
