+++ apps/main/src/llamalend/features/borrow/queries/borrow-price-impact.query.ts
@@ -1,7 +1,7 @@
+import { getLlamaMarket } from '@/llamalend/llama.utils'
 import { LendMarketTemplate } from '@curvefi/llamalend-api/lib/lendMarkets'
 import { queryFactory, rootKeys } from '@ui-kit/lib/model'
-import type { BorrowFormQuery, BorrowFormQueryParams } from '../borrow.types'
-import { getLlamaMarket } from '../llama.util'
+import type { BorrowFormQuery, BorrowFormQueryParams } from '../types'
 import { borrowExpectedCollateralQueryKey } from './borrow-expected-collateral.query'
 import { borrowQueryValidationSuite } from './borrow.validation'
