+++ apps/main/src/llamalend/features/borrow/queries/borrow.validation.ts
@@ -2,8 +2,8 @@ import { enforce, group, test } from 'vest'
 import { createValidationSuite, type FieldsOf } from '@ui-kit/lib'
 import { chainValidationGroup } from '@ui-kit/lib/model/query/chain-validation'
 import { llamaApiValidationGroup } from '@ui-kit/lib/model/query/curve-api-validation'
-import { type BorrowForm, type BorrowFormQueryParams } from '../borrow.types'
-import { BORROW_PRESET_RANGES } from '../llama.util'
+import { BORROW_PRESET_RANGES } from '../constants'
+import { type BorrowForm, type BorrowFormQueryParams } from '../types'
 
 const validateUserBorrowed = (userBorrowed: number | null | undefined) =>
   test('userBorrowed', 'Borrow amount must be a non-negative number', () => {
