+++ apps/main/src/llamalend/features/borrow/queries/create-loan.mutation.ts
@@ -1,5 +1,6 @@
 import { useCallback } from 'react'
 import { useAccount, useConfig } from 'wagmi'
+import { getLlamaMarket } from '@/llamalend/llama.utils'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { LendMarketTemplate } from '@curvefi/llamalend-api/lib/lendMarkets'
 import { useMutation } from '@tanstack/react-query'
@@ -10,9 +11,8 @@ import { t } from '@ui-kit/lib/i18n'
 import { Address } from '@ui-kit/utils'
 import { waitForTransactionReceipt } from '@wagmi/core'
 import { getBalanceQueryKey } from '@wagmi/core/query'
-import type { BorrowForm, BorrowFormQuery } from '../borrow.types'
-import { getLlamaMarket } from '../llama.util'
 import { borrowExpectedCollateralQueryKey } from '../queries/borrow-expected-collateral.query'
+import type { BorrowForm, BorrowFormQuery } from '../types'
 import { borrowFormValidationSuite } from './borrow.validation'
 import { userBalancesQueryKey } from './user-balances.query'
