+++ apps/main/src/llamalend/widgets/tooltips/RewardTooltipItems.tsx
@@ -1,4 +1,4 @@
-import { formatPercent } from '@/llamalend/utils'
+import { formatPercent } from '@/llamalend/format.utils'
 import ArrowOutwardIcon from '@mui/icons-material/ArrowOutward'
 import { Stack } from '@mui/material'
 import Link from '@mui/material/Link'
